refactor: move components to checklist folder
